@@ -7,7 +7,7 @@ name = "pymimir-rl"
 version = "0.1.0"
 description = "Reinforcement Learning (RL) package for Mimir."
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.11"
 license = "GPL-3.0-or-later"
 authors = [
   { name = "Simon Stahlberg", email = "simon.stahlberg@gmail.com" }
@@ -17,9 +17,6 @@ classifiers = [
   "Intended Audience :: Developers",
   "Intended Audience :: Science/Research",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
